renoved one of vlients types
