+++ src/Byndyusoft.ApiClient/Byndyusoft.ApiClient.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
 		<TargetFramework>netstandard2.0</TargetFramework>
@@ -13,6 +13,9 @@
 
 	<ItemGroup>
 		<PackageReference Include="Byndyusoft.Net.Http" Version="0.4.4" />
+		<PackageReference Include="Byndyusoft.Net.Http.Json" Version="0.3.2" />
+		<PackageReference Include="Byndyusoft.Net.Http.MessagePack" Version="0.6.0" />
+		<PackageReference Include="Byndyusoft.Net.Http.ProtoBuf" Version="0.5.0" />
 		<PackageReference Include="Microsoft.Extensions.Options" Version="8.0.2" />
 		<PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
 	</ItemGroup>
