+++ tests/Byndyusoft.ApiClient.Tests.csproj
@@ -28,6 +28,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\example\Byndyusoft.ApiClient.Example.csproj" />
     <ProjectReference Include="..\src\Byndyusoft.ApiClient\Byndyusoft.ApiClient.csproj" />
   </ItemGroup>
